fix route protection
